+++ src/Compiler/Compiler/AbstractCompiler.php
@@ -82,7 +82,7 @@ abstract class AbstractCompiler implements CompilerInterface
         $providers = $this->engine->getProviders();
         $providers['uiControl'] = new ObjectType($actualClass ?? 'Nette\Application\UI\Control');
         $providers['uiPresenter'] = new ObjectType($actualClass ?? 'Nette\Application\UI\Presenter');
-        $providers['snippetDriver'] = TypeCombinator::addNull(new ObjectType('Nette\Bridges\ApplicationLatte\SnippetDriver'));
+        $providers['snippetDriver'] = new ObjectType('Nette\Bridges\ApplicationLatte\SnippetDriver');
         $providers['uiNonce'] = TypeCombinator::addNull(new StringType());
         $providers['formsStack'] = new ArrayType(new IntegerType(), new ObjectType('Nette\Forms\Container'));