+++ tests/Rule/LatteTemplatesRule/PresenterWithoutModule/Fixtures/SnippetsPresenter.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\PHPStanLatte\Tests\Rule\LatteTemplatesRule\PresenterWithoutModule\Fixtures;
+
+use Nette\Application\UI\Presenter;
+
+final class SnippetsPresenter extends Presenter
+{
+    public function renderDefault(): void
+    {
+    }
+}